+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.7.0] - 2020-07-22
 ### Added
 - `<ImageSlider>` - - LinkTo prop added
+- `<List>` - - component added for home & marketplace list view
+- `<TabPanels>` - - component added with custom Tab & Tabs style option
 
 ## [1.6.0] - 2020-07-15
 ### Added